Move ToolConfiguration wholesale to it's own file
